use decimal to validate amounts

+++ hive_attention_tokens/chain/transactions/validators/definitions.py
@@ -1,11 +1,12 @@
 """Defines all transaction types and validation rules."""
+from decimal import Decimal
 
 """ trans_type, to_account, token, amount """
 AIRDROP = [
     [str, 3],
     [str, 16],
     [str, 12],
-    [float, None, 3]
+    [Decimal, None, 3]
 ]
 
 """ trans_type, from_account, to_account, token, amount """
@@ -15,5 +16,5 @@ TRANSFER = [
     [str, 16],
     [str, 16],
     [str, 12],
-    [float, None, 3]
+    [Decimal, None, 3]
 ]
